Fix incomplete release notes in the first release by this action (and improve logging) (#5)
* Update logging to improve debugging

* Change release note titles from "##" to "#"

Which is in line with the existing style, see release v2.9.0 and before 
at: https://github.com/simple-icons/simple-icons/releases

* Avoid using `break` in double for-loop + switch-statement construction

* Fix function to check if a PR is a release

This fixes the issue that caused the first release created by this
Action to be incorrect.

Previously, the check to see if a PR is the previous release was based
on the `head` of that PR (that is, the branch from which we want to
merge). If that was "develop" it was considered a release. However, if a
contributor created a Pull Request from their `develop` branch, the
Action would think that was a release.

To fix this, the check has been updated to see if the `base` of the PR
(that is, the branch into which we want to merge) is master. Since the
`base` is always the repo where the action is running, we only have to
check into which branch it is merging.