Fix earlier changes included in release notes
Due to a misunderstanding of how the PRs are sorted, the last release 
notes [1] included too many changes. This was caused by the fact that 
some of the PRs included in the release before it were updated after 
that release was merged (e.g. the brach was removed).

Now, when all changes of all PRs updated after the last release are 
gathered, the time at which the PR of that file was merged is compared 
to the time at which the previous release was merged, and the file is 
included in the next release accordingly.

For this comparison, Moment.JS was used as the GitHub API returns the 
merged time as a string rather then a standard Unix epoch timestamp.

--
1.  https://github.com/simple-icons/simple-icons/pull/3019