Fix changes to brand colour missing from release notes
The regular epxression for detecting changes in the datafile only 
included a case for changes to the source URL, but was missing a case 
for changes to the colour value. This has now been fixed.

In addition, there are two minor fixes to the @actions/github mock:

- The patch for the modified Postman SVG is now actually for Postman
- The patch for the datafile in PR 509 no longer contains a duplicate 
for AddThis (to avoid possibly confusing release notes problems in the 
future)