Fix updated brands listed multiple times in release notes (#14)
Fix the bug where an update to a brand may appear multiple times
in the release notes. This was caused by the fact that the list of
updates was not checked against itself.

In addition, the checking for double changes was previously done
through the file path. This doesn't work when an icon was updated
in both the SVG and the data file. Therefore, this has been updated
to be based on the name (instead of path).