Build to .cjs file
... to be able to run the action in the GitHub Actions context.

Because the source is now ESM (signified by `"type": "module"` in
package.json), Node and GitHub actions expect also the lib/ file to be
an ESM module, but NCC still compiles to a CommonJS module.

An alternative solution would be to put a separate package.json file in
the lib/ directory.