Remove moment.js
Remove the dependency on moment.js, which was used only to check if
files are merged before or after the previous release Pull Request.
This behaviour can easily be replicated using JavaScript's Date [1], as
per [2]. My main motivation for doing this now is the project status
notice on Moment.js' website [3]. This states that the project is in
maintainance mode and therefore has limited support.

--
1. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date
2. https://stackoverflow.com/a/493018
3. https://momentjs.com/docs/#/-project-status/